clean up create_webite
